Updating when we use Key attributes and changing tests to match. Moving dbo into the SqlServer provider. Also refactoring Sqlite tests to use a common base class.
