+++ test/EntityFramework.SqlServer.Design.FunctionalTests/ReverseEngineering/ExpectedResults/E2E_UseAttributesInsteadOfFluentApi/OneToManyDependent.expected
@@ -9,9 +9,7 @@ namespace E2ETest.Namespace
 {
     public class OneToManyDependent
     {
-        [Key]
         public int OneToManyDependentID1 { get; set; }
-        [Key]
         public int OneToManyDependentID2 { get; set; }
         public int? OneToManyDependentFK1 { get; set; }
         public int? OneToManyDependentFK2 { get; set; }
