+++ test/EntityFramework.Sqlite.Design.FunctionalTests/ReverseEngineering/Expected/UseAttributesInsteadOfFluentApi/OneToMany/OneToManyPrincipal.expected
@@ -14,9 +14,7 @@ namespace E2E.Sqlite
             OneToManyDependent = new HashSet<OneToManyDependent>();
         }
 
-        [Key]
         public long OneToManyPrincipalID1 { get; set; }
-        [Key]
         public long OneToManyPrincipalID2 { get; set; }
         [Required]
         public string Other { get; set; }
