+++ test/EntityFramework.Sqlite.Design.FunctionalTests/ReverseEngineering/Expected/UseAttributesInsteadOfFluentApi/OneToOne/ModelContext.expected
@@ -7,7 +7,7 @@ namespace E2E.Sqlite
     {
         protected override void OnConfiguring(DbContextOptionsBuilder options)
         {
-            options.UseSqlite(@"Data Source=OneToOne.db;Cache=Private");
+            options.UseSqlite(@"Data Source=OneToOneAttributes.db;Cache=Private");
         }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
